backend/requirements.txt
# FastAPI Backend Dependencies for Phase II Todo Application
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0  # For SQLite async support

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# CORS
fastapi-cors==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
